feat/#507: 중복 코드 제거 및 reissueToken 메소드 리팩토링
