+++ BE/exceed/src/main/java/com/gaebaljip/exceed/application/service/auth/AuthService.java
@@ -1,6 +1,8 @@
 package com.gaebaljip.exceed.application.service.auth;
 
-import com.gaebaljip.exceed.common.EatCeedStaticMessage;
+import com.gaebaljip.exceed.adapter.out.redis.RedisAdapter;
+import com.gaebaljip.exceed.common.exception.auth.NotFoundRefreshTokenException;
+import com.gaebaljip.exceed.common.security.domain.JwtResolver;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -9,6 +11,7 @@ import com.gaebaljip.exceed.adapter.in.auth.request.LoginRequest;
 import com.gaebaljip.exceed.application.domain.member.MemberEntity;
 import com.gaebaljip.exceed.application.port.in.auth.AuthUsecase;
 import com.gaebaljip.exceed.application.port.out.member.MemberPort;
+import com.gaebaljip.exceed.common.EatCeedStaticMessage;
 import com.gaebaljip.exceed.common.dto.HttpRequestDTO;
 import com.gaebaljip.exceed.common.dto.TokenDTO;
 import com.gaebaljip.exceed.common.exception.auth.PasswordMismatchException;
@@ -25,6 +28,8 @@ public class AuthService implements AuthUsecase {
     private final MemberPort memberPort;
     private final BCryptPasswordEncoder bCryptPasswordEncoder;
     private final JwtManager jwtManager;
+    private final JwtResolver jwtResolver;
+    private final RedisAdapter redisAdapter;
 
     @Override
     public TokenDTO execute(LoginRequest request) {
@@ -37,7 +42,9 @@ public class AuthService implements AuthUsecase {
                         .accessToken(jwtManager.generateAccessToken(member.getId()))
                         .refreshToken(jwtManager.generateRefreshToken(member.getId()))
                         .build();
-        jwtManager.saveRefreshToken(EatCeedStaticMessage.REDIS_REFRESH_TOKEN_KEY + member.getId().toString(), tokenDTO.refreshToken());
+        jwtManager.saveRefreshToken(
+                EatCeedStaticMessage.REDIS_REFRESH_TOKEN_KEY + member.getId().toString(),
+                tokenDTO.refreshToken());
         return tokenDTO;
     }
 
@@ -45,7 +52,25 @@ public class AuthService implements AuthUsecase {
     public TokenDTO reIssueToken(
             String accessToken, String refreshToken, HttpRequestDTO requestDTO) {
         if (jwtManager.validateRefreshToken(refreshToken, requestDTO)) {
-            return jwtManager.reissueToken(accessToken);
+            return reissueToken(accessToken);
+        }
+        throw InvalidJwtException.EXECPTION;
+    }
+
+    private TokenDTO reissueToken(String requestAccessToken) {
+        String accessToken = jwtResolver.extractToken(requestAccessToken);
+        String accessTokenMemberId = jwtResolver.getMemberIdFromToken(accessToken);
+        String refreshToken =
+                redisAdapter
+                        .query(EatCeedStaticMessage.REDIS_REFRESH_TOKEN_KEY + accessTokenMemberId)
+                        .orElseThrow(() -> NotFoundRefreshTokenException.EXECPTION);
+        String refreshTokenMemberId = jwtResolver.getMemberIdFromToken(refreshToken);
+
+        if (accessTokenMemberId.equals(refreshTokenMemberId)) {
+            return TokenDTO.builder()
+                    .accessToken(jwtManager.generateAccessToken(Long.parseLong(accessTokenMemberId)))
+                    .refreshToken(jwtManager.generateRefreshToken(Long.parseLong(refreshTokenMemberId)))
+                    .build();
         }
         throw InvalidJwtException.EXECPTION;
     }
